+++ timex/doc.go
@@ -25,7 +25,7 @@ import (
 
 // PackageVersion returns the version of the version package.
 func PackageVersion() version.Version {
-	return version.New(3, 1, 0)
+	return version.New(3, 2, 0)
 }
 
 // EOF